more refactor and readme
