+++ src/book.py
@@ -16,12 +16,12 @@
 
 import base64
 import collections
-import csv
 import datetime
 import decimal
 import hashlib
 import hmac
 import json
+import os
 import re
 import time
 from collections import defaultdict
@@ -38,15 +38,8 @@ import transaction as tr
 from core import kraken_asset_map
 from database import set_price_db
 from exchanges.base import ExchangeReader
-from exchanges.binance import BinanceReader
 from exchanges.bitget_api import BitgetApiReader
-from exchanges.bitpanda import BitpandaReader
-from exchanges.bitunix import BitunixReader
-from exchanges.coinbase import CoinbaseReader
-from exchanges.coinbase_pro import CoinbaseProReader
-from exchanges.custom_eur import CustomEurReader
-from exchanges.kraken import KrakenReader
-from exchanges.pionex import PionexReader
+from exchanges.registry import detect_exchange_reader
 from price_data import PriceData
 
 log = log_config.getLogger(__name__)
@@ -58,36 +51,6 @@ class MissingOperation(NamedTuple):
 
     def repr(self) -> str:
         return f"- {self.platform}: {self.operation}"
-
-
-def create_exchange_reader(exchange_name: str) -> Optional[ExchangeReader]:
-    """Create an exchange reader instance based on the exchange name."""
-    reader_map = {
-        "binance": BinanceReader,
-        "binance_v2": BinanceReader,
-        "binance_v3": BinanceReader,
-        "coinbase": CoinbaseReader,
-        "coinbase_v2": CoinbaseReader,
-        "coinbase_v3": CoinbaseReader,
-        "coinbase_v4": CoinbaseReader,
-        "coinbase_pro": CoinbaseProReader,
-        "kraken_ledgers_old": KrakenReader,
-        "kraken_ledgers": KrakenReader,
-        "kraken_trades": KrakenReader,
-        "bitpanda_pro_trades": BitpandaReader,
-        "bitpanda": BitpandaReader,
-        "bitunix": BitunixReader,
-        "pionex_deposit_withdraw": PionexReader,
-        "pionex_trading": PionexReader,
-        "pionex_staking": PionexReader,
-        "pionex_others": PionexReader,
-        "custom_eur": CustomEurReader,
-    }
-
-    reader_class = reader_map.get(exchange_name)
-    if reader_class:
-        return reader_class()
-    return None
 
 
 class Book:
@@ -176,295 +139,37 @@ class Book:
             if op is not None:
                 self._append_operation(op)
 
-    def import_bitget_api_records(self) -> None:
-        log.info("Importing Bitget records from API for tax year %s.", config.TAX_YEAR)
-        year_start = datetime.datetime(
-            config.TAX_YEAR, 1, 1, tzinfo=datetime.timezone.utc
-        )
-        year_end = datetime.datetime(
-            config.TAX_YEAR, 12, 31, 23, 59, 59, 999000, tzinfo=datetime.timezone.utc
-        )
+    def _get_bitget_record_types_from_env(self) -> Optional[list[str]]:
+        """Parse selected Bitget API record groups from environment.
+
+        The env var `BITGET_API_RECORD_TYPES` accepts a comma separated list,
+        for example: `spot,future,margin,p2p`.
+        """
+        configured = os.environ.get("BITGET_API_RECORD_TYPES", "").strip()
+        if not configured:
+            return None
+        return [value.strip() for value in configured.split(",") if value.strip()]
+
+    def import_api_records(self) -> None:
+        """Import records from configured exchange APIs."""
+        if not (
+            config.BITGET_API_KEY
+            and config.BITGET_API_SECRET
+            and config.BITGET_API_PASSPHRASE
+        ):
+            return
+
+        record_types = self._get_bitget_record_types_from_env()
         bitget_reader = BitgetApiReader()
-        bitget_reader.import_spot_records(
-            self, int(year_start.timestamp() * 1000), int(year_end.timestamp() * 1000)
-        )
-        bitget_reader.import_future_records(
-            self, int(year_start.timestamp() * 1000), int(year_end.timestamp() * 1000)
-        )
-        bitget_reader.import_margin_records(
-            self, int(year_start.timestamp() * 1000), int(year_end.timestamp() * 1000)
-        )
-        bitget_reader.import_p2p_records(
-            self, int(year_start.timestamp() * 1000), int(year_end.timestamp() * 1000)
+        log.info("Importing Bitget records from API for tax year %s.", config.TAX_YEAR)
+        bitget_reader.import_tax_year_records(
+            self,
+            config.TAX_YEAR,
+            record_types=record_types,
         )
 
     def detect_exchange(self, file_path: Path) -> Optional[ExchangeReader]:
-        if file_path.suffix == ".csv":
-
-            expected_header_row = {
-                "binance": 1,
-                "binance_v2": 1,
-                "binance_v3": 1,
-                "coinbase": 1,
-                "coinbase_v2": 1,
-                "coinbase_v3": 1,
-                "coinbase_v4": 4,
-                "coinbase_pro": 1,
-                "kraken_ledgers_old": 1,
-                "kraken_ledgers": 1,
-                "kraken_trades": 1,
-                "bitpanda_pro_trades": 4,
-                "bitpanda": 7,
-                "bitunix": 1,
-                "pionex_deposit_withdraw": 1,
-                "pionex_trading": 1,
-                "pionex_staking": 1,
-                "pionex_others": 1,
-                "custom_eur": 1,
-            }
-
-            expected_headers = {
-                "binance": [
-                    "UTC_Time",
-                    "Account",
-                    "Operation",
-                    "Coin",
-                    "Change",
-                    "Remark",
-                ],
-                "binance_v2": [
-                    "User_ID",
-                    "UTC_Time",
-                    "Account",
-                    "Operation",
-                    "Coin",
-                    "Change",
-                    "Remark",
-                ],
-                "binance_v3": [
-                    "\ufeffUser ID",
-                    "Time",
-                    "Account",
-                    "Operation",
-                    "Coin",
-                    "Change",
-                    "Remark",
-                ],
-                "coinbase": [
-                    "You can use this transaction report to inform your "
-                    "likely tax obligations. For US customers, Sells, "
-                    "Converts, and Rewards Income, and Coinbase Earn "
-                    "transactions are taxable events. For final tax "
-                    "obligations, please consult your tax advisor."
-                ],
-                "coinbase_v2": [
-                    "You can use this transaction report to inform your "
-                    "likely tax obligations. For US customers, Sells, "
-                    "Converts, Rewards Income, Coinbase Earn "
-                    "transactions, and Donations are taxable events. "
-                    "For final tax obligations, please consult your tax advisor."
-                ],
-                "coinbase_v3": [
-                    "You can use this transaction report to inform your "
-                    "likely tax obligations. For US customers, Sells, "
-                    "Converts, Rewards Income, Learning Rewards, "
-                    "and Donations are taxable events. "
-                    "For final tax obligations, please consult your tax advisor."
-                ],
-                "coinbase_v4": [
-                    "ID",
-                    "Timestamp",
-                    "Transaction Type",
-                    "Asset",
-                    "Quantity Transacted",
-                    "Price Currency",
-                    "Price at Transaction",
-                    "Subtotal",
-                    "Total (inclusive of fees and/or spread)",
-                    "Fees and/or Spread",
-                    "Notes",
-                ],
-                "coinbase_pro": [
-                    "portfolio",
-                    "trade id",
-                    "product",
-                    "side",
-                    "created at",
-                    "size",
-                    "size unit",
-                    "price",
-                    "fee",
-                    "total",
-                    "price/fee/total unit",
-                ],
-                "kraken_ledgers_old": [
-                    "txid",
-                    "refid",
-                    "time",
-                    "type",
-                    "aclass",
-                    "asset",
-                    "amount",
-                    "fee",
-                    "balance",
-                ],
-                "kraken_ledgers": [
-                    "txid",
-                    "refid",
-                    "time",
-                    "type",
-                    "subtype",
-                    "aclass",
-                    "asset",
-                    "amount",
-                    "fee",
-                    "balance",
-                ],
-                "kraken_trades": [
-                    "txid",
-                    "ordertxid",
-                    "pair",
-                    "time",
-                    "type",
-                    "ordertype",
-                    "price",
-                    "cost",
-                    "fee",
-                    "vol",
-                    "margin",
-                    "misc",
-                    "ledgers",
-                ],
-                "bitpanda_pro_trades": [
-                    "Order ID",
-                    "Trade ID",
-                    "Type",
-                    "Market",
-                    "Amount",
-                    "Amount Currency",
-                    "Price",
-                    "Price Currency",
-                    "Fee",
-                    "Fee Currency",
-                    "Time (UTC)",
-                ],
-                "bitpanda": [
-                    "Transaction ID",
-                    "Timestamp",
-                    "Transaction Type",
-                    "In/Out",
-                    "Amount Fiat",
-                    "Fiat",
-                    "Amount Asset",
-                    "Asset",
-                    "Asset market price",
-                    "Asset market price currency",
-                    "Asset class",
-                    "Product ID",
-                    "Fee",
-                    "Fee asset",
-                    "Spread",
-                    "Spread Currency",
-                ],
-                "bitunix": [
-                    "Date (UTC)",
-                    "Label",
-                    "Outgoing Asset",
-                    "Outgoing Amount",
-                    "Incoming Asset",
-                    "Incoming Amount",
-                    "Fee Asset",
-                    "Fee Amount",
-                    "Trx. ID",
-                    "Comment",
-                ],
-                "pionex_deposit_withdraw": [
-                    "date(UTC+0)",
-                    "tx_type",
-                    "amount",
-                    "coin",
-                    "network",
-                    "txid",
-                    "fee",
-                ],
-                "pionex_trading": [
-                    "date(UTC+0)",
-                    "executed_qty",
-                    "amount",
-                    "price",
-                    "side",
-                    "symbol",
-                    "fee",
-                    "fee_coin",
-                    "market_type",
-                    "tax_id",
-                ],
-                "pionex_staking": [
-                    "date(UTC+0)",
-                    "Received Quantity",
-                    "Received Currency",
-                    "Sent Quantity",
-                    "Sent Currency",
-                    "tag",
-                ],
-                "pionex_others": [
-                    "date(UTC+0)",
-                    "coin",
-                    "amount",
-                    "tag",
-                    "comment",
-                ],
-                "custom_eur": [
-                    "Type",
-                    "Buy Quantity",
-                    "Buy Asset",
-                    "Buy Value in EUR",
-                    "Sell Quantity",
-                    "Sell Asset",
-                    "Sell Value in EUR",
-                    "Fee Quantity",
-                    "Fee Asset",
-                    "Fee Value in EUR",
-                    "Wallet",
-                    "Timestamp UTC",
-                    "Note",
-                ],
-            }
-
-            # Special handling for Pionex which has multiple file types
-            filename = file_path.name
-            pionex_files = {
-                "deposit-withdraw.csv": "pionex_deposit_withdraw",
-                "trading.csv": "pionex_trading",
-                "staking.csv": "pionex_staking",
-                "others.csv": "pionex_others",
-            }
-            if filename in pionex_files:
-                exchange_type = pionex_files[filename]
-                with open(file_path, encoding="utf8") as f:
-                    reader = csv.reader(f)
-                    expected = expected_headers[exchange_type]
-                    header = next(reader, None)
-                    if header == expected:
-                        return create_exchange_reader(exchange_type)
-
-            with open(file_path, encoding="utf8") as f:
-                reader = csv.reader(f)
-                # check all potential headers at their expected header row
-                for exchange, expected in expected_headers.items():
-                    # Skip Pionex entries as they're handled above
-                    if exchange.startswith("pionex_"):
-                        continue
-                    header_row_num = expected_header_row[exchange]
-                    # iterate since header row may appear earlier
-                    for _ in range(header_row_num):
-                        header = next(reader, None)
-                        if header == expected:
-                            return create_exchange_reader(exchange)
-                    # rewind the file after each header check
-                    f.seek(0)
-
-        return None
+        return detect_exchange_reader(file_path)
 
     def resolve_deposits(self) -> None:
         """Match withdrawals to deposits.
